third_party/blink/web_tests/editing/pasteboard/drag-and-drop-image-contenteditable-expected.txt

CONSOLE WARNING: Listener added for a synchronous 'DOMNodeInserted' DOM Mutation Event. This event type is deprecated (https://w3c.github.io/uievents/#legacy-event-types) and work is underway to remove it from this browser. Usage of this event listener will cause performance issues today, and represents a risk of future incompatibility. Consider using MutationObserver instead.
This test verify dragging an image in the page keeps its attributes. To test manually, drag the image and drop it in the rect with a blue border.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS target.getElementsByTagName("img").length is 1
PASS target.getElementsByTagName("img")[0].attributes.length is 5
PASS target.getElementsByTagName("img")[0].id is "source"
PASS target.getElementsByTagName("img")[0].alt is "alt-text"
PASS target.getElementsByTagName("img")[0].attributes["itemprop"].value is "somedata"
PASS target.getElementsByTagName("img")[0].attributes["arbitrary"].value is "green"
PASS successfullyParsed is true

TEST COMPLETE
